cleaner Settingsprovider
